Добавляет загрузку фотографии
