+++ js/loading-photo.js
@@ -1,7 +1,7 @@
 const FILE_TYPES = ['jpg', 'jpeg', 'png'];
 const fileChooser = document.querySelector('.img-upload__input');
 const preview = document.querySelector('.img-upload__preview img');
-// const effectsImg = document.querySelectorAll('.effects__label span');
+const effectsImg = document.querySelectorAll('.effects__label span');
 
 fileChooser.addEventListener('change', () => {
   const file = fileChooser.files[0];
@@ -9,6 +9,8 @@ fileChooser.addEventListener('change', () => {
   const matches = FILE_TYPES.some((it) => fileName.endsWith(it));
   if (matches) {
     preview.src = URL.createObjectURL(file);
-    // effectsImg.src = URL.createObjectURL(file);
   }
+  effectsImg.forEach((item) => {
+    item.style.backgroundImage = `url(${URL.createObjectURL(file)})`;
+  });
 });
